+++ checkout-ui-custom/src/controller/DeliverController.js
@@ -20,6 +20,7 @@ import {
   clearRicaFields,
   customShippingDataIsValid,
   parseAttribute,
+  populateAddresses,
   populateDeliveryError,
   populateRicaFields,
   populateTVFields,
@@ -128,6 +129,8 @@ const DeliverController = (() => {
         hasFurnMixed: state.hasFurnMixed,
       })
     );
+
+    populateAddresses();
 
     if (state.hasFurn) {
       $('#shipping-data:not(.has-furniture)').addClass('has-furniture');
